requirements.txt
app-version==1.0.1
arabic-reshaper==2.1.3
asgiref==3.3.1
astroid==2.4.2
auth==0.5.3
autopep8==1.5.4
backcall==0.2.0
bleach==4.1.0
blinker==1.4
Brotli==1.0.9
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
convertdate==2.3.2
cryptography==3.3.1
cssselect2==0.4.1
cycler==0.11.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.13
django-allauth==0.44.0
django-appconf==1.0.5
django-avatar==5.0.0
django-bower==5.2.0
django-cors-headers==3.6.0
django-datetimepicker==3.14
django-debug-toolbar==3.2
django-easy-pdf==0.1.1
django-filter==21.1
django-jsonfield==1.4.1
django-markdown-filter==0.0.1
django-markdownify==0.9.0
django-mdeditor==0.1.18
django-notify-x==0.1.9
django-pandas==0.6.6
django-taggit==2.0.0
djangorestframework==3.12.2
dnspython==1.16.0
eventlet==0.30.0
falcon==2.0.0
flake8==3.8.4
fonttools==4.28.1
future==0.18.2
greenlet==0.4.17
gunicorn==20.0.4
hijri-converter==2.2.2
holidays==0.11.3.1
html5lib==1.1
idna==2.10
importlib-metadata==4.8.2
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.4.3
Markdown==3.3.6
markdown2==2.4.1
matplotlib==3.5.0
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.22.1
numpy==1.19.4
oauthlib==3.1.0
packaging==21.3
pandas==1.1.5
parso==0.7.1
pdfkit==1.0.0
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pydyf==0.1.2
pyflakes==2.2.0
Pygments==2.7.3
PyJWT==2.0.0
pylint==2.6.0
PyMeeus==0.5.11
pymongo==3.11.2
pyparsing==3.0.6
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python3-openid==3.2.0
pytz==2020.4
reportlab==3.6.2
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
setuptools-scm==6.3.2
signals==0.0.2
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.2
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==4.0.1
unicode==2.8
Unidecode==1.3.2
urllib3==1.26.2
validate-email==1.3
wcwidth==0.2.5
WeasyPrint==0.42.3
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xhtml2pdf==0.2.5
yapf==0.30.0
zipp==3.4.0
zopfli==0.1.9
